Bug fix: field `zcj' may be float

+++ vruc_score_bot_dotnet_core/GradeResult.cs
@@ -6,7 +6,7 @@
         public string kcname { get; set; }
         public int? xf { get; set; }
         public double? jd { get; set; }
-        public int? zcj { get; set; }
+        public double? zcj { get; set; }
         public string cjxm1 { get; set; }
         public string cjxm3 { get; set; }
     }
